Propagating Working on Worm, not working on Omozra
Made Propagate Transfrom recursive, now it sets the objects to dirty and iterates, not 100% code, it works for the worm so i assumed i'm in the good direction.

The collision boxes of the box don't move, the skeleton does, that's some important info.

The recursivity sometimes don't hit the objects before the skeleton and doesn't move them (Omozra problem)

We're near but something is missing / not going well